refactor(jsii): refactor jsii folder
