+++ src/jsii/index.ts
@@ -1,3 +1,3 @@
-export * from './npm';
+export * from './config/npm';
 export * from './project';
 export * from './typescript';
