+++ src/jsii/project.ts
@@ -1,15 +1,15 @@
 import { cdk } from 'projen';
-import { CommitLintConfigBase } from '../base/commitlint';
-import { DevContainerConfigBase } from '../base/devcontainer';
-import { EsLintConfigBase } from '../base/eslint';
-import { GitHubConfigBase } from '../base/github';
-import { HuskyConfigBase } from '../base/husky';
-import { GitConfigBase } from '../base/git';
-import { JestConfigBase } from '../base/jest';
-import { BaseOptions } from '../base/options';
-import { PrettierConfigBase } from '../base/prettier';
-import { VsCodeConfigBase } from '../base/vscode';
-import { NpmConfigJsii } from './npm';
+import { CommitLintConfigBase } from '../base/config/commitlint';
+import { DevContainerConfigBase } from '../base/config/devcontainer';
+import { EsLintConfigBase } from '../base/config/eslint';
+import { GitHubConfigBase } from '../base/config/github';
+import { HuskyConfigBase } from '../base/config/husky';
+import { GitConfigBase } from '../base/config/git';
+import { JestConfigBase } from '../base/config/jest';
+import { BaseOptions } from '../base/config/options';
+import { PrettierConfigBase } from '../base/config/prettier';
+import { VsCodeConfigBase } from '../base/config/vscode';
+import { NpmConfigJsii } from './config/npm';
 import { TypeScriptConfigJsii } from './typescript';
 import { registerConfig } from '../utils';
 import { IProjectKind, ProjectKind } from '../types/project';
